Add files via upload
Find the winner on counting from 1 to n on a given elements. an element is termed as loser or out or removed if it is having a count  divisor of 4 or giving a remainder on modulo 4 division. return the element which is the winner. 

sample input : 
elements = { a,b,c,d,e,f,g,h,i,j,k}